build(vite): remove YAML-related settings and serve JSONs
We have removed YAML-related parser and configurations before, so we
remove YAML settings here.

And `__PROJECT_ROOT__` has not been used yet, so we remove it.

Note that the default value of `server.fs.strict` is true, which does
not allow to import other modules outside the allow list. However,
adding `./config.json?import` or `./config.json` manually does not work.
So we disable strict importing.