chore: use new types and add `parseVars()`
This is a continuation of bc6d090e96cd009dbfa79592e9d3ecab818786cd. It
uses new types and rewrites some logic in the code of components and
config utilites. And it adds `getVarsSubstitutions()` and `parseVars()`
which can parse `$`-style raw strings (i.e. `homework.savePath`) to
the target ones.