feat(route): add i18n support
This commit adds some routes and component parameters to the existed
endpoints. Adding "lang=<language identifier>" to the URL query string
can change the viewing language.

But note that the standalone components (a.k.a. those Svelte components
outside `src/routes` in the old version) don't support i18n now, because
they lack a parameter to change the language. It will be fixed in future
commits.