docs(readme): move logger from to-do list to feature list
The logger runs to some extent. However, the logger is unable to capture
errors from the client-side (`hooks.client.ts` missing) and assignments
which do not meet the rules are not to be logged, for the only rule
checker (`FilenameCheck`) is running on the client-side, so there is no
any request sent to the server-side if assignment files fail the
validation. We need more enhancement here, but this issue can be marked
as completed, *temporarily*.

Close #10